requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# OpenAI
openai==1.12.0

# Twilio
twilio==9.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Vector Database (supporting multiple backends)
pinecone-client==3.0.0
weaviate-client==4.4.0
chromadb==0.4.22

# RAG & Embeddings
langchain==0.1.6
langchain-openai==0.0.5
sentence-transformers==2.2.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
unstructured==0.11.6
beautifulsoup4==4.12.2
langdetect==1.0.9
markdownify==0.11.6
playwright==1.40.0

# Audio Processing
pydub==0.25.1
websockets==12.0
python-socketio[asyncio]==5.11.0

# Task Queue
celery==5.3.6
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Sentiment Analysis & NLP
textblob==0.17.1
vaderSentiment==3.3.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.26.0
aiofiles==23.2.1
reportlab==4.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Security
cryptography==42.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

